requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
mlflow==2.8.0
MetaTrader5==5.0.45
python-telegram-bot==20.6
pandas==2.1.3
numpy==1.26.2
pydantic==2.4.2
pytest==7.4.3